make an option for using long double for calculations
The floating point typedef used for calculations is named realt.
The previous name fp was too cryptic, and `real' is already used in std
namespace. Use realt only for storing data, data operations, functions
and fitting. Use double where it does not affect fitting.

+++ src/wxgui/datatable.cpp
@@ -97,8 +97,10 @@ public:
               char buffer[128];
               sprintf(buffer,
                       "M=M+1, X[%d]=%.12g, Y[%d]=%.12g, S[%d]=%.12g, A[%d]=%d",
-                      row, last_point_.x, row, last_point_.y,
-                      row, last_point_.sigma, row, last_point_.is_active);
+                      row, (double) last_point_.x,
+                      row, (double) last_point_.y,
+                      row, (double) last_point_.sigma,
+                      row, last_point_.is_active);
               change_value(buffer);
               if (!instant_update_)
                   local_points_.push_back(last_point_);
